macd_param_analysis/requirements.txt
sshtunnel
sqlalchemy
pandas
yfinance
vectorbt
psycopg2
asyncpg
paramiko<4
pyyaml

polars[all]
seaborn
